Refactor README to link to LICENSE.md for clarity. Update EnhancedNPCBehaviors to streamline gang war news announcements. Modify NPC emotional state updates to consider recent memories. Adjust potion purchase logic in MagicShopLocation for type consistency. Simplify player location updates in NewsLocation and enhance input handling in PrisonWalkLocation. Improve combat calculations in AdvancedCombatEngine and CombatEngine with explicit type casting. Add ConsoleColor overloads in TerminalEmulator for legacy compatibility and remove redundant methods from legacy compatibility files. These changes aim to enhance code clarity, maintainability, and user experience.

+++ Scripts/Locations/MagicShopLocation.cs
@@ -493,7 +493,7 @@ public partial class MagicShopLocation : BaseLocation
         // Calculate potion price: level × 5
         int potionPrice = player.Level * GameConfig.HealingPotionLevelMultiplier;
         int maxPotionsCanBuy = (int)(player.Gold / potionPrice);
-        int maxPotionsCanCarry = GameConfig.MaxHealingPotions - player.Healing;
+        int maxPotionsCanCarry = GameConfig.MaxHealingPotions - (int)player.Healing;
         int maxPotions = Math.Min(maxPotionsCanBuy, maxPotionsCanCarry);
         
         if (player.Gold < potionPrice)
